+++ app/pages/home/MissionSection.jsx
