+++ constants/Images.js
@@ -13,6 +13,7 @@ import Quiz from "@assets/icons/quiz.svg";
 import Walk from "@assets/icons/walk.svg";
 import Diet from "@assets/icons/diet.svg";
 import Target from "@assets/icons/missionTarget.svg";
+import Target2 from "@assets/icons/missionTarget2.svg";
 import SuccessDiet from "@assets/icons/successDiet.svg";
 import SuccessWalk from "@assets/icons/successWalk.svg";
 
@@ -50,6 +51,7 @@ export const Images = {
   Walk,
   Diet,
   Target,
+  Target2,
   SuccessDiet,
   SuccessWalk,
 };
